+++ src/main/resources/templates/user/reset.html
@@ -1,0 +1,26 @@
+<!-- Reset Password Modal -->
+<div class="modal fade modal-dialog-scrollable" id="modal-resetpw" tabindex="-1" aria-labelledby="modal-label-resetpw"
+     aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="modal-label-resetpw">Enter Your Email</h5>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div class="modal-body">
+                <form action="emailResetPW" method="post">
+                    <div class="mb-3">
+                        <label for="login-email-resetpw" class="form-label">Email address</label>
+                        <input type="email" class="form-control" name="email" id="login-email-resetpw"
+                               required>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                        <button type="submit" class="btn btn-primary">Submit</button>
+                    </div>
+                </form>
+            </div>
+
+        </div>
+    </div>
+</div>
