+++ src/test/java/com/websummarizer/Web/Summarizer/checkRequestTokenTest.java
@@ -1,6 +1,7 @@
 package com.websummarizer.Web.Summarizer;
 
 import com.websummarizer.Web.Summarizer.bart.Bart;
+import com.websummarizer.Web.Summarizer.controller.PasswordResetController;
 import com.websummarizer.Web.Summarizer.controller.WebController;
 import com.websummarizer.Web.Summarizer.model.User;
 import com.websummarizer.Web.Summarizer.repo.UserRepo;
@@ -17,7 +18,7 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 
-@WebMvcTest(WebController.class)
+@WebMvcTest(PasswordResetController.class)
 @AutoConfigureMockMvc(addFilters = false)   // Disallow SecurityConfig from messing up the tests
 
 public class checkRequestTokenTest {
